pyproject.toml
[tool.poetry]
name = "drf-cw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"
coverage = "^7.5.4"


[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
mypy = "^1.10.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
